refactor(webui): move app/components to top-level components/ per AGENTS.md
