+++ services/web/app/pages/request-demo-page.tsx
@@ -5,7 +5,7 @@ import { Input } from '@tale/ui/input';
 import { Textarea } from '@tale/ui/textarea';
 import { useForm } from 'react-hook-form';
 
-import { FormCard } from '@/app/components/blocks/form-card';
+import { FormCard } from '@/components/blocks/form-card';
 import {
   REQUEST_DEMO_INTERESTS,
   type RequestDemoInput,
