+++ services/web/components/blocks/hardware-spec-value.tsx
