+++ services/web/components/blocks/hardware-specs.ts
