+++ services/web/components/blocks/hero-headline.tsx
@@ -1,8 +1,8 @@
 import { Button } from '@tale/ui/button';
 import { motion, useReducedMotion } from 'framer-motion';
 
-import { LocalizedLink } from '@/app/components/layout/localized-link';
-import { SiteContainer } from '@/app/components/layout/site-container';
+import { LocalizedLink } from '@/components/layout/localized-link';
+import { SiteContainer } from '@/components/layout/site-container';
 import { useT } from '@/lib/i18n/client';
 
 const easeOut = [0.22, 1, 0.36, 1] as const;
