+++ services/web/components/icons/microsoft-icon.tsx
