+++ services/web/components/layout/site-footer.tsx
@@ -6,12 +6,12 @@ import {
 import { Link } from '@tanstack/react-router';
 import type { ReactNode } from 'react';
 
-import { GithubIcon } from '@/app/components/icons/github-icon';
-import { ExternalLink } from '@/app/components/layout/external-link';
+import { GithubIcon } from '@/components/icons/github-icon';
+import { ExternalLink } from '@/components/layout/external-link';
 import {
   LocalizedLink,
   type LocalizedRoutePath,
-} from '@/app/components/layout/localized-link';
+} from '@/components/layout/localized-link';
 import { useT } from '@/lib/i18n/client';
 import { useCurrentLocale } from '@/lib/i18n/use-current-locale';
 import type { LegalSlug } from '@/lib/legal/slugs';
